Complete methods for getting account balance

+++ tenmo-server/src/main/java/com/techelevator/tenmo/controller/AccountController.java
@@ -19,7 +19,7 @@ public class AccountController {
 
     @PreAuthorize("isAuthenticated()")
     @RequestMapping (path = "/accounts/{id}", method = RequestMethod.GET)
-    public BigDecimal getUserBalance(@PathVariable int id) {
+    public BigDecimal getUserBalance(@PathVariable Long id) {
         return accountDao.getUserBalance(id);
     }
